Rebase themes changing (#321)
* Added theme color dial in the frontend, added new themeAdd view in views.py
for userprofile and also add the respective url for it. Added the themeadd
function in the api.js of the frontend. There seem to be some problem posting
my newly added theme to the backend.

* Implemented theme add, theme delete and theme update functions in views.py of user profile

* theme creation, deletion and update functionality

* fixed erros for themeadd

* Implemented Theme Favorite and Unfavorite functions

* Updated theme favorite feature

* Added theme color dial in the frontend, added new themeAdd view in views.py
for userprofile and also add the respective url for it. Added the themeadd
function in the api.js of the frontend. There seem to be some problem posting
my newly added theme to the backend.

* Implemented theme add, theme delete and theme update functions in views.py of user profile

* theme creation, deletion and update functionality

* fixed erros for themeadd

* Implemented Theme Favorite and Unfavorite functions

* rebased themechanging with master

* fixed bugs

* reference error keep occuring

* error unfixed

* added tool tip tip indicate colors

* fixed themechanging after theme addition and theme edit

* removed node modules and package lock

* changed sidebar icon in post, the color of the icon follows the secondary font color

* limited theme name to 20 chars, modified theme delete function

themes currently being used and starred by other users will not be deleted, the theme author will
be set null

on theme change/theme unstar, if the last user deselects or unstar the theme, the theme will be deleted

* merged edit and creation modals
used deepcopy with lodash.clonedeep for state updates

* added animations, modeified layout of user profiles main page

* removed console logs and fixed warnings

* used serializers for error handling
fixed height for user profile theme row

* fixed the bug with deleting the selected theme, updated the theme delete method in backend to set author to none upon theme deletion, modified the load additional themes method

* removed sudos from init script

* fixed two-way favorting for theme and users in many to many relation (in user profile model signals)

Co-authored-by: Rex <rixin@ymail.com>